Finish Encode/Decode destination functions move from base58 to key_io.
