+++ src/core_write.cpp
@@ -5,7 +5,7 @@
 
 #include "core_io.h"
 
-#include "base58.h"
+#include "key_io.h"
 #include "primitives/transaction.h"
 #include "script/script.h"
 #include "script/standard.h"
