+++ src/evo/deterministicmns.cpp
@@ -5,11 +5,11 @@
 
 #include "evo/deterministicmns.h"
 
-#include "base58.h"
 #include "chainparams.h"
 #include "consensus/upgrades.h"
 #include "core_io.h"
 #include "evo/specialtx.h"
+#include "key_io.h"
 #include "guiinterface.h"
 #include "masternode.h" // for MasternodeCollateralMinConf
 #include "masternodeman.h" // for mnodeman (!TODO: remove)
