+++ src/httprpc.cpp
@@ -5,10 +5,10 @@
 
 #include "httprpc.h"
 
-#include "base58.h"
 #include "chainparams.h"
 #include "crypto/hmac_sha256.h"
 #include "httpserver.h"
+#include "key_io.h"
 #include "rpc/protocol.h"
 #include "rpc/server.h"
 #include "random.h"
