+++ src/masternode.h
@@ -6,7 +6,7 @@
 #ifndef MASTERNODE_H
 #define MASTERNODE_H
 
-#include "base58.h"
+#include "key_io.h"
 #include "key.h"
 #include "messagesigner.h"
 #include "net.h"
