+++ src/messagesigner.cpp
@@ -3,7 +3,6 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "base58.h"
 #include "hash.h"
 #include "key_io.h"
 #include "messagesigner.h"
