+++ src/qt/walletmodel.h
@@ -14,6 +14,7 @@
 #include "interfaces/wallet.h"
 
 #include "key.h"
+#include "key_io.h"
 #include "operationresult.h"
 #include "support/allocators/zeroafterfree.h"
 #include "pairresult.h"
