+++ src/rpc/budget.cpp
@@ -5,10 +5,11 @@
 
 #include "activemasternode.h"
 #include "chainparams.h"
-#include "db.h"
-#include "init.h"
 #include "budget/budgetmanager.h"
+#include "db.h"
 #include "evo/deterministicmns.h"
+#include "init.h"
+#include "key_io.h"
 #include "masternode-payments.h"
 #include "masternode-sync.h"
 #include "masternodeconfig.h"
