+++ src/rpc/misc.cpp
@@ -5,10 +5,10 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "base58.h"
 #include "clientversion.h"
 #include "httpserver.h"
 #include "init.h"
+#include "key_io.h"
 #include "sapling/key_io_sapling.h"
 #include "masternode-sync.h"
 #include "net.h"
