+++ src/rpc/rawtransaction.cpp
@@ -5,7 +5,6 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "base58.h"
 #include "core_io.h"
 #include "evo/specialtx.h"
 #include "evo/providertx.h"
