+++ src/test/bloom_tests.cpp
@@ -5,9 +5,9 @@
 
 #include "bloom.h"
 
-#include "base58.h"
 #include "clientversion.h"
 #include "key.h"
+#include "key_io.h"
 #include "merkleblock.h"
 #include "random.h"
 #include "serialize.h"
