+++ src/test/rpc_tests.cpp
@@ -6,7 +6,7 @@
 #include "rpc/server.h"
 #include "rpc/client.h"
 
-#include "base58.h"
+#include "key_io.h"
 #include "netbase.h"
 #include "util/system.h"
