+++ src/wallet/wallet.h
@@ -10,13 +10,13 @@
 
 #include "addressbook.h"
 #include "amount.h"
-#include "base58.h"
 #include "consensus/tx_verify.h"
 #include "consensus/validation.h"
 #include "crypter.h"
 #include "destination_io.h"
 #include "kernel.h"
 #include "key.h"
+#include "key_io.h"
 #include "keystore.h"
 #include "pairresult.h"
 #include "policy/feerate.h"
